requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
discord.py==1.7.2
google-auth==1.30.1
google-auth-oauthlib==0.4.4
gspread==3.7.0
httplib2==0.19.1
idna==2.10
multidict==5.1.0
oauth2client==4.1.3
oauthlib==3.1.1
Pillow==8.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-http-client==3.3.2
reportlab==3.5.68
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sendgrid==6.7.1
six==1.16.0
starkbank-ecdsa==1.1.1
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
